Load app earlier.
Previously the Backbone app had been firing on window.load with some odd logic to pick it up if window.load had already fired. This moves it up to document.ready under the assumption that the previous incarnation was due to various requirements imposed by the require.js method of loading the app. So far this seems to be working well.