Check each role for presence of capability
This adds a check of the `editor` role capabilities and splits out the addition of the capabilities into separate functions.

If the `editor` role somehow gets reset to base capabilities, it will never receive the brightcove caps again since the `administrator` role still contains the correct capabilities.

Adding a check of the `editor` role, and adding the role capabilities individually ensures that each role will get reassigned necessary capabilities if they are ever lost for any reason.